Move make_retty_image to images.py, add some doc strings.
