+++ pocs/utils/images/conversions.py
@@ -95,50 +95,3 @@ def process_cr2(cr2_fname, to_fits=True, fits_headers={}, solve=False, make_pret
     """
 
 
-def make_pretty_image(fname, timeout=15, **kwargs):
-    """ Make a pretty image
-
-    This calls out to an external script which will try to extract the JPG directly from the CR2 file,
-    otherwise will do an actual conversion
-
-    Notes:
-        See `$POCS/scripts/cr2_to_jpg.sh`
-
-    Arguments:
-        fname {str} -- Name of CR2 file
-        **kwargs {dict} -- Additional arguments to be passed to external script
-
-    Keyword Arguments:
-        timeout {number} -- Process timeout (default: {15})
-
-    Returns:
-        str -- Filename of image that was created
-
-    """
-    assert os.path.exists(fname), warn("File doesn't exist, can't make pretty: {}".format(fname))
-
-    verbose = kwargs.get('verbose', False)
-
-    title = '{} {}'.format(kwargs.get('title', ''), current_time().isot)
-
-    solve_field = "{}/scripts/cr2_to_jpg.sh".format(os.getenv('POCS'), '/var/panoptes/POCS')
-    cmd = [solve_field, fname, title]
-
-    if kwargs.get('primary', False):
-        cmd.append('link')
-
-    if verbose:
-        print(cmd)
-
-    try:
-        proc = subprocess.Popen(cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-        if verbose:
-            print(proc)
-    except OSError as e:
-        raise error.InvalidCommand("Can't send command to gphoto2. {} \t {}".format(e, run_cmd))
-    except ValueError as e:
-        raise error.InvalidCommand("Bad parameters to gphoto2. {} \t {}".format(e, run_cmd))
-    except Exception as e:
-        raise error.PanError("Timeout on plate solving: {}".format(e))
-
-    return fname.replace('cr2', 'jpg')
